+++ packages/components/molecules/InfoBox/types/index.ts
@@ -1,3 +1,6 @@
 export interface InfoBoxProps {
   label: string;
+  testProp?: string;
+  coolProp?: number;
+  isInformative?: boolean;
 }