+++ packages/types/index.ts
@@ -1,27 +1,27 @@
-import type { ButtonProps } from '@no-gravity-elements/button/types';
-import type { IconProps } from '@no-gravity-elements/icon/types';
-import type { InfoBoxProps } from '@no-gravity-elements/info-box/types';
-import type { LinkProps } from '@no-gravity-elements/link/types';
-import type { TypographyProps } from '@no-gravity-elements/typography/types';
+// import type { ButtonProps } from '@no-gravity-elements/button/types';
+// import type { IconProps } from '@no-gravity-elements/icon/types';
+// import type { InfoBoxProps } from '@no-gravity-elements/info-box/types';
+// import type { LinkProps } from '@no-gravity-elements/link/types';
+// import type { TypographyProps } from '@no-gravity-elements/typography/types';
 
-type ReactWebComponent =
-  | (Partial<HTMLElement> &
-      Partial<React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement>, HTMLElement>>)
-  // @ts-ignore
-  | { children?: React.ReactNode };
+// type ReactWebComponent =
+//   | (Partial<HTMLElement> &
+//       Partial<React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement>, HTMLElement>>)
+//   // @ts-ignore
+//   | { children?: React.ReactNode };
 
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      'nge-button': ButtonProps & ReactWebComponent;
-      'nge-card': ReactWebComponent;
-      'nge-icon': IconProps & ReactWebComponent;
-      'nge-link': LinkProps & ReactWebComponent;
-      'nge-typography': TypographyProps & ReactWebComponent;
-      'nge-info-box': InfoBoxProps & ReactWebComponent;
-    }
-  }
-}
+// declare global {
+//   namespace JSX {
+//     interface IntrinsicElements {
+//       'nge-button': ButtonProps & ReactWebComponent;
+//       'nge-card': ReactWebComponent;
+//       'nge-icon': IconProps & ReactWebComponent;
+//       'nge-link': LinkProps & ReactWebComponent;
+//       'nge-typography': TypographyProps & ReactWebComponent;
+//       'nge-info-box': InfoBoxProps & ReactWebComponent;
+//     }
+//   }
+// }
 
 export enum TypographySizes {
   h1 = 'h1',
@@ -64,3 +64,5 @@ export enum ButtonTypes {
   ghost = 'ghost',
   disabled = 'disabled',
 }
+
+export default {};