+++ packages/components/atoms/Typography/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.22](https://github.com/no-gravity-company/no-gravity-elements/compare/@no-gravity-elements/typography@1.0.21...@no-gravity-elements/typography@1.0.22) (2024-03-30)
+
+**Note:** Version bump only for package @no-gravity-elements/typography
+
 ## [1.0.21](https://github.com/no-gravity-company/no-gravity-elements/compare/@no-gravity-elements/typography@1.0.16...@no-gravity-elements/typography@1.0.21) (2024-03-29)
 
 ### Bug Fixes